Repackage hash-related APIs from Guava into proofs module:
Hash-related APIs, as of Guava 23, are in Beta, which means we,
as a library, cannot expose them to our users. Therefore, we have
a couple of choices: (a) repackage the whole thing; (b) roll out
our own hash classes; (c) use some other library with a stable API;
(d) repackage the needed classes manually.

(a) isn't great because we _still expose_ the Beta APIs, and it will
complicate build configuration, make builds take longer time
and make Guava update more difficult (or even undesirable).

(b) is a good option, but requires to implement and test everything
ourselves.

(c) does not appear in a quick search.

(d) slightly increases our maintenance costs for we now include
the code. On the other hand, we fully control it and can make
any adjustments to suit our needs. It also seems unlikely
we will need to update the classes in `com.exonum.binding.hash`
manually with a new upstream release.