Add a test leaking a native peer:
The problem with most index constructors is that they first
create a native peer, and them perform some checks, e.g., in
ProofListIndex constructor:

super(nativeCreate(checkIndexName(name),
  view.getViewNativeHandle()), name, view, serializer);

Therefore, the parent receives a native handle, and then checks
some of the other arguments (serializer, in this case). If these checks
fail, we get a leak of a native peer.

A possible solution is to simplify error checking with static factory
methods, that do not have such a restriction that the first call
in the method must be a call to `super`. To make it better, we may
also wrap the code in try/finally, reliably destroying the native peer.