Add a converter between transactions and their messages [ECR-760]:
It makes it easy to decouple a transaction (its data) from
the transaction message, and eases the creation of transactions
in tests. Transactions then have to create the corresponding messages
themselves.

However, in real application such interface will not work as is,
because a transaction message cannot be fully created only
from the transaction parameters, a couple of extra parameters
are needed:
  - a network id,
  - a protocol version,
  - the message signature.
You may say that these parameters are meta-data about transaction,
because they are not a part of its body, they do not affect
the code executing it.