Make destruction of native objects implicit [ECR-595]: (#244)
Add Cleaner — a class managing lifecycle of native proxies [ECR-595].

Cleaner accepts CleanActions and performs them in a reversed order
of the registration when closed. Cleaners are created and closed
by the framework, providing a controlled context to the user code,
so that all user-created proxies are definitely destroyed.

Cleaners have description so that an origin of the instance can be identified.

There are two types of proxies at the moment:
* Proxies that must be explicitly destroyed (e.g., MemoryDb).
* Proxies that are automatically destroyed (e.g., View, ListIndexProxy).
The latter do not implement CloseableNativeProxy interface so that
there is no #close method available to the end-users and no risk 
that they misuse it. Instead they register a clean action freeing 
the corresponding native peer (usually, a ProxyDestructor).

ECR-1287 ECR-595 #done Resolved via PR #244