Log state hashes of QaService [ECR-1371]: (#262)
The state hashes are required for QA purposes to test the integration
with the Exonum Core. See the logging requirements on the wiki:
https://wiki.bf.local/display/EJB/QA+Service+Requirements

Also, remove the default implementation of Schema#getStateHashes,
so that users make a deliberate choice to return an empty list
(i.e., not return it by ignorance).