Accept Exonum JSON messages in cryptocurrency service [ECR-1755]:
JSON messages are used to pass transaction parameters,
the message header and signature in human-readable format. There is
usually no need to complicate things with numerous conversions
to/from JSON in the light client and the backend.

Binary messages are no longer created in the transactions,
but are passed as a single constructor parameter. This is a proper
behaviour, enabling verification of digital signatures
of the messages.

Also, PublicKey is no longer used in proto messages as
it complicates the client code significantly with no much benefit.