Include paths to shared libraries in artefacts [ECR-1793]: (#336)
Add paths to runtime libraries required to run the application or load 
the library during image generation. This allows the linker to find
the required shared libraries on a local machine without extra
environment variables (= LD_LIBRARY_PATH on Linux/Mac),
some of which cannot be easily passed with enabled SIP on Mac.

The paths can still be overridden with LD_LIBRARY_PATH.

See the docs in "tests_profile" for more details.