Add TransactionExecutionException [ECR-1958]: (#392)
Such exception is thrown to indicate that the transaction cannot be
executed and any changes must be rolled back. Its error code
is included in the blockchain and can be requested by a client.

This exception corresponds to ExecutionError in the framework
native API: https://docs.rs/exonum/0.8.0/exonum/blockchain/struct.ExecutionError.html

Also, add a method to create a transaction mock throwing 
TransactionExecutionException to be able to test the new behaviour
in the native proxy code.