Separate Travis CI jobs (#405)
Moved various code checks that do not depend on
the toolchain version to a separate script to speed up the build:
 - Cargo fmt (requires the nightly compiler)
 - Cargo audit
 - Clippy
 - Maven versions check

The script uses the CHECK_RUST environment variable to decide
if it shall perform the code checks or run all tests. Currently the code
checks are performed with JDK 8.