Build only specific pushed branches [ECR-2386]:
As Travis builds both branches and PRs by default, it will create two
builds for a single PR that uses a branch pushed to a repo.
To avoid that (and longer queues for other builds), specify
protected-only to be built.

One can still create WIP PRs to get them built on Travis if they need.

See https://docs.travis-ci.com/user/customizing-the-build/#safelisting-or-blocklisting-branches