Update TransactionProxy#execute to use transaction status from Java [ECR-1956] (#482)
The result of execution of a transaction is treated in a little different way now: any JNI error or Java throwable that is not an instance of the TransactionExecutionException is considered an unexpected exception and causes panic. TransactionExecutionException and its descendants are converted into Errors with their descriptions and error codes. 

Former behavior: JNI errors and anything other than lang.java.Exception and its descendants considered unexpected exception --> panic.