Update indexes to accept proto messages [ECR-2340]
Make them automatically create reflective serializers of
protobuf messages in their static factory methods.

Note 1: ProofMapIndexProxy does not get such a method —
it has a certain restriction on keys, which is unlikely
to hold for an arbitrary Protocol Buffers message.

Note 2: dependency on protobuf cannot be optional
because the compiler has to load the classes appearing
in parameter lists to determine which of the overloaded
methods to call. It can be optional if we _require_
the clients to add it in "provided" scope.

Also, avoid overloads by method reference. Because of some
compiler bug in JDK 8, a call like
IndexConstructors.from(ListIndexProxy::newInstance)
cannot be resolved as ambiguous. The same code compiles fine on 11.