Migrate Java part to Exonum 0.10 [ECR-2400] (#574)
Update the Java part to Exonum 0.10 and its new message format,
that does not require executable transactions to keep a copy of
the original transaction message they were created from.
The classes supporting the previous format (Exonum 0.1-0.9)
are removed.

QA service and cryptocurrency-demo service are updated to use
the new APIs. Cryptocurrency now accepts the transactions through
a system endpoint instead of a custom one in the service REST API:
`<public api IP address and port>/api/explorer/v1/transactions`
This and a new message format breaks the cryptocurrency frontend,
see ECR-2752/#630 for a fix.

Notes:
  - Uses a nightly Rust version, which has some known problems
  on Travis CI
  - Cryptocurrency frontend is broken, see ECR-2752/#630

Co-authored-by: Denis Makarov <denm2000@gmail.com>
Signed-off-by: Dmitry Timofeev <dmytro.tymofeev@xdev.re>