Add native part of time oracle service: [ECR-2815]
Add the time service factory to the list of supported built-in
service factories. Also, add a means to check if the time service
is currently activated. It is done by checking if the time service
is enabled in the configuration file — to be replaced with
a more appropriate mechanism when runtimes/dynamic services land.

Signed-off-by: Dmitry Timofeev <dmytro.tymofeev@xdev.re>