Configure services as artifacts
Configure projects-services to package the classes in a fat JAR with
meta data describing them. The service dependencies that are
currently the framework dependencies are marked as 'provided', however,
such decision is to be discussed (Are we certain we want to make
these dependencies guaranteed to be available with the framework?).

ServiceModules are also renamed for clarity.