Make the ServiceRuntime actually thread-safe [ECR-3065]:
Make the ServiceRuntime actually thread-safe without reliance
on ServiceLoader thread-safety, because ServiceRuntime is a Singleton
and *may* be injected into other objects. That does not happen
currently; nor the ServiceRuntime is used in a multi-threaded
context, hence this feature is provided to avoid possible
synchronization errors if it does happen in the future (e.g.,
if in dynamic services the operations on ServiceRuntime are invoked
on different threads).

The other runtime-specific binding are put in a private module,
i.e., they are inaccessible to satisfy bindings from other
modules or from injector directly.