Use a PluginManager tailored for JARs [ECR-3066]:
Override the operations in DefaultPluginManager that produce
services that support both types of plugins (ZIP and JARs) or that
do not make sense with JARs or our use cases.

That allows to reduce the number of I/O operations during plugin
loading (as of 2.6.x, ZIP has higher priority than JAR) and the number
of redundant log messages.

If no longer needed, JarPluginManager class can be replaced with
a DefaultPluginManager.