Check no copies of application libraries in artifact [ECR-3046]:
Check that no copies of application libraries are included in the
service artifact during artifact loading. It helps to prevent
errors in build definitions when a dependency does not have
'provided' scope set, as recently discovered ECR-3091.

The present version cannot check if time-oracle is included
because it is not available on the core classpath.