Move log4j configuration for native ITs
Move the log4j configuration for native ITs to native ITs
so that it is (a) easily discoverable; (b) can be changed with no
recompilation of Java `fakes` module.

Also, pass the path to the Log4j configuration file to the
service runtime bootstrap test to remove the error message from logs.