Skip unit tests without resource-manager in the dev environment: (#920)
Run the unit tests of the Rust library without resource-manager
on CI only. Earlier (in ancient times) it was not possible 
because `cargo test`  used to overwrite the library *with* 
rm previously built by `cargo build`, which is required for 
subsequent Java ITs.

The current order of test actions on CI for core:
 - default-test
 - test-rust-lib
 - test-rust-lib-no-resource-manager <-- Used to cause problems at this position
 - integration-test

These tests now won't run on `mvn test`, but still will on
`./run_all_tests.sh` or `mvn test -P ci-build`.