Rename InternalServerError and make unchecked [ECR-3258] (#947)
Rename InternalServerError to InvalidTransactionException (as it
more precisely reflects the _most likely_ exception from Node#submit)
and make it unchecked (as _any_ exception is unrecoverable).

There is, arguably, the same issue that the exception type does not
always work — if ApiSender#send fails, than the exception type
is misleading. That is to be addressed in ECR-3190.