Add TestKit#getSnapshot [ECR-3216] (#972)
Add TestKit#getSnapshot for most common cases when only
several snapshots are created and can be removed when
testkit is closed.

Make TestKit#withSnapshot accept Consumer<Snapshot> as a parameter,
and add TestKit#applySnapshot. These methods remain for (rare) usages
that require a large number of snapshots or native proxies and cannot
rely on de-allocation on Testkit#close.