Add CI for the `merkledb` branch:
Exonum has to be cloned manually because Cargo does not
allow specifying a git dependency on a module which is
not the root project.

RocksDB currently has to be linked statically
(as there are some problems with dynamic linking)
and always built (as the PPA we use does not have
a static library, see ECR-3277). See the ECR-3311
for more details.