Support conversion of Fork into Patch [ECR-3364] (#1026)
Add Fork into Patch conversion so that the changes can be applied
to the MerkleDB database. This conversion consumes the original
fork object, therefore, it is needed to support 'native peer ownership
transfer' from Java to native. The exact mechanics of how it would
work in native are to be determined, see #1024 and ECR-3330