Relax index modification restrictions [ECR-3322]:
Allow to iterate over one index and modify any other independent
index as this is determined to be safe with MerkleDB.
Modification of the index during iteration over it (even if it
happens through separate proxies) is still prohibited and must
be reliably detected. As indexes are de-duplicated now, no extra
effort is required to ensure such reliable detection (except the tests).