Update service runtime to newer Exonum revision 4 [ECR-3802]: (#1219)
Updates the service runtime to the new Exonum revision (exonum/exonum@2bbea66).

Extracted the connection of service APIs into RuntimeTransport.
Also fixed the documentation of Server#start clarifying that start
happens asynchronously; and modified the method to return a Future
to enable the clients to know when the op is complete.