Integrate blockchain data [ECR-4175] (#1393)
Integrated BlockchainData into service APIs: Service, TransactionContext, Node, TestKit.

BlockchainData is an object providing access to all categories of persistent data stored in the database: the executing service data, other services’ data, Exonum Core data.

BlockchainData is a service-specific object — it remembers the service to which it is provided and allows modification only to the data of that service. The service data is automatically isolated via namespaces, with a service name followed by a dot as a prefix.

Node is made a service-specific interface — and now provides BlockchainData.

NodeProxy is no longer a Node, as it works with Snapshots.

Removed AbstractService#createDataSchema — just use the schema constructor/factory method.