RoFork -> RoErasedAccess [ECR-4176] (#1400)
Replace RoFork with RoErasedAccess, as native always operates with ErasedAccesses,
and in Java it is not known what the BlockchainData uses under the hood.

Fixes the previous incorrect assumption that BlockchainData was based on Fork. It was actually benign as both ReadonlyFork (and new RoErasedAccess) and Snapshot are based on the same type of native peer (ErasedAccess).