Move messages to a separate module [ECR-4199] (#1421)
Moved Exonum-core messages to a separate module. The module is still a child project of ejb and is versioned as ejb; but has a group id to match its root package. See the reasoning in Jira.

Also move potentially-needed by blockchain and/or node client apps proto definitions to common.