Fix the test of Ed25519 operation with the system sodium:
The Ed25519CryptoFunction used to initialize an instance of
itself in a static initializer, making it impossible to
instantiate it with a different library loading mode (as
the library can only be loaded once into a JVM process).

This PR moves the instance initialization to CryptoFunctions,
where it happens lazily (just as previously) in a holder class.